Change the layout of command one-liner descriptions.
Instead of having the syntax on the leftmost 40 columns and the
description on the rightmost 40 columns, causing many commands to span
more than one line total, we now display the syntax on a whole line, and
the description on the next one, indented by a tab.

This should hopefully improve readability (especially for complex
syntax lines).