Workaround non-unicity of (zone, backend domain id) tuple with views.
When views are enabled, make the UeberBackend lookup handle only query
backends for zones which variant match that of the request. Without that,
there is a risk of returning results for a different variant if it exists
in a different backend, with the same domain id as the queried variant
zone.

Signed-off-by: Miod Vallat <miod.vallat@powerdns.com>